Proxy structure improvements (#134)
Closes #125 

## Objectives

* Enable CoCo to contribute in a directed and safe way
* Enable Registry to contribute in a directed and safe way
* Decouple domain logic from specific API implementation (GraphQL)
* Improve maintainability

## Structure

In essence the new source structure looks like this:

```
proxy/
├── Cargo.lock
├── Cargo.toml
├── README.md
├── rust-toolchain
├── src
│   ├── coco.rs
│   ├── env.rs
│   ├── error.rs
│   ├── graphql
│   │   ├── api.rs
│   │   ├── error.rs
│   │   ├── project.rs
│   │   └── schema.rs
│   ├── graphql.rs
│   ├── lib.rs
│   ├── main.rs
│   └── registry.rs
└── tests
    ├── common
    │   └── mod.rs
    ├── graphql_mutation.rs
    └── graphql_query.rs

```

This will reap the following benefits:
* decoupling API from core domain
* single entry point for protocol teams
* Adhere to 2018 [Path clarity](https://doc.rust-lang.org/edition-guide/rust-2018/module-system/path-clarity.html)
* Test high-level API through [integration tests](https://doc.rust-lang.org/book/ch11-03-test-organization.html#integration-tests)

***
* Rename server to api
* Rearrange graphql code including the api
* Move API setup into respective module
* Nove schema entities out of top-level
* Remove leftover
* Move registry out of graphql module
* Remove graphql leftovers out of registry
* Rearrange schema
* Move coco out of graphql
* Use anynomuous struct for variant
* Move coco logic out of graphql schema
* Move graphql conversions out of coco
* Move API tests over to integration section
* Please the clippy
* Please cloudhead
* Fix typo
