+++ proxy/src/graphql/api.rs
@@ -3,11 +3,25 @@ use warp::filters::BoxedFilter;
 use warp::http::Response;
 use warp::Filter;
 
-use crate::schema::{Context, Schema};
+use super::schema::Context;
 
 /// Runs the warp server with the given schema and context.
-pub fn run(schema: Schema, context: Context) {
-    let routes = make_graphql_filter("graphql", schema, context)
+pub fn run(
+    dummy_repo_path: String,
+    librad_paths: librad::paths::Paths,
+    registry_client: radicle_registry_client::Client,
+) {
+    let context = Context::new(dummy_repo_path, librad_paths, registry_client);
+    let schema = super::schema::create();
+
+    let index = warp::path::end().map(|| {
+        Response::builder()
+            .header("content-type", "text/html")
+            .body(juniper::graphiql::graphiql_source("/graphql"))
+    });
+
+    let routes = index
+        .or(make_graphql_filter("graphql", schema, context))
         .with(
             warp::cors()
                 .allow_any_origin()
